requirements.txt
streamlit
ultralytics
opencv-python-headless
yt-dlp
numpy
torch
ffmpeg-python